refactor: 지역명 상수화
