+++ src/utils/constance.ts
@@ -1,3 +1,22 @@
 export const INIT_LOCATION = { latitude: 37.5665, longitude: 126.978 };
 export const FINE_DUST = '미세먼지';
 export const ULTRA_FINE_DUST = '초미세먼지';
+export const CITY_GROUP = [
+  { cityName: '서울', cityNumber: 0 },
+  { cityName: '부산', cityNumber: 1 },
+  { cityName: '대구', cityNumber: 2 },
+  { cityName: '인천', cityNumber: 3 },
+  { cityName: '광주', cityNumber: 4 },
+  { cityName: '대전', cityNumber: 5 },
+  { cityName: '울산', cityNumber: 6 },
+  { cityName: '경기', cityNumber: 7 },
+  { cityName: '강원', cityNumber: 8 },
+  { cityName: '충북', cityNumber: 9 },
+  { cityName: '충남', cityNumber: 10 },
+  { cityName: '전북', cityNumber: 11 },
+  { cityName: '전남', cityNumber: 12 },
+  { cityName: '경북', cityNumber: 13 },
+  { cityName: '경남', cityNumber: 14 },
+  { cityName: '제주', cityNumber: 15 },
+  { cityName: '세종', cityNumber: 16 },
+];
